Use formatting options in date/time rendering
