+++ src/main/java/org/rumbledb/runtime/functions/datetime/TemporalFormattingSupport.java
@@ -4,9 +4,9 @@ import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.util.Locale;
 
-import org.rumbledb.runtime.functions.base.formatting.NumericPictureParser;
-import org.rumbledb.runtime.functions.base.formatting.timezone.TimezoneNameContext;
-import org.rumbledb.runtime.functions.base.formatting.timezone.TimezoneNameRegistry;
+import org.rumbledb.runtime.functions.util.formatting.NumericPictureParser;
+import org.rumbledb.runtime.functions.util.formatting.timezone.TimezoneNameContext;
+import org.rumbledb.runtime.functions.util.formatting.timezone.TimezoneNameRegistry;
 
 final class TemporalFormattingSupport {
